src/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.2
flask-jwt-extended==4.7.1
flask-limiter==4.1.1
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.3
nltk==3.8.1
psycopg2-binary==2.9.9
requests==2.32.3
google-generativeai==0.3.2
gunicorn==21.2.0
pytest==8.3.4
pytest-cov==6.0.0
